# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Message {
  id: number;
  fanEmail: string;
  fanName: string;
  message: string;
  amountPaid: number;
  txRef: string;
  status: string;
  /** @nullable */
  reply?: string | null;
  createdAt: string;
}

export interface MessageInput {
  fanEmail: string;
  fanName: string;
  message: string;
  amountPaid: number;
  txRef: string;
}

export interface MessageUpdate {
  status?: string;
  reply?: string;
}

export interface CallBooking {
  id: number;
  fanEmail: string;
  fanName: string;
  preferredDate: string;
  durationMinutes: number;
  amountPaid: number;
  txRef: string;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CallBookingInput {
  fanEmail: string;
  fanName: string;
  preferredDate: string;
  durationMinutes: number;
  amountPaid: number;
  txRef: string;
  notes?: string;
}

export interface CallBookingUpdate {
  status?: string;
  notes?: string;
}

export interface ContentRequest {
  id: number;
  fanEmail: string;
  fanName: string;
  requestType: string;
  description: string;
  amountPaid: number;
  txRef: string;
  status: string;
  createdAt: string;
}

export interface ContentRequestInput {
  fanEmail: string;
  fanName: string;
  requestType: string;
  description: string;
  amountPaid: number;
  txRef: string;
}

export interface Tip {
  id: number;
  fanEmail: string;
  fanName: string;
  amount: number;
  txRef: string;
  /** @nullable */
  message: string | null;
  createdAt: string;
}

export interface TipInput {
  fanEmail: string;
  fanName: string;
  amount: number;
  txRef: string;
  message?: string;
}

export interface Stats {
  totalMessages: number;
  totalCalls: number;
  totalRequests: number;
  totalTips: number;
  totalRevenue: number;
}

export type ListMessagesParams = {
status?: ListMessagesStatus;
};

export type ListMessagesStatus = typeof ListMessagesStatus[keyof typeof ListMessagesStatus];


export const ListMessagesStatus = {
  pending: 'pending',
  paid: 'paid',
  replied: 'replied',
} as const;