# lib/api-zod/src/generated/types/callBooking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */

export interface CallBooking {
  id: number;
  fanEmail: string;
  fanName: string;
  preferredDate: string;
  durationMinutes: number;
  amountPaid: number;
  txRef: string;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}