# lib/api-zod/src/generated/types/contentRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentRequestInput {
  fanEmail: string;
  fanName: string;
  requestType: string;
  description: string;
  amountPaid: number;
  txRef: string;
}