# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  fanEmail: string;
  fanName: string;
  message: string;
  amountPaid: number;
  txRef: string;
  status: string;
  /** @nullable */
  reply?: string | null;
  createdAt: string;
}