# lib/api-zod/src/generated/types/messageUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageUpdate {
  status?: string;
  reply?: string;
}