# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hannah Brooks Fan Site API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalMessages: number;
  totalCalls: number;
  totalRequests: number;
  totalTips: number;
  totalRevenue: number;
}